backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
google-generativeai==0.3.1
sentence-transformers==2.2.2
bertopic==0.15.0
python-dotenv==1.0.0